Made it so it didn't crash on floats

+++ src/c2.rs
@@ -1,18 +1,18 @@
 // A function to get a number e.g. age or adding numbers
-pub fn get_int() -> i64 {
+pub fn get_int() -> f64 {
     let mut line = String::new();
     loop {
         std::io::stdin()
             .read_line(&mut line)
             .expect("Failed to read line");
-        let line: i64 = match line.trim().parse() {
+        let line: f64 = match line.trim().parse() {
             Ok(num) => num,
             Err(_) => {
                 println!("That's not a number");
                 continue;
             }
         };
-        return line;
+        return line as f64;
     }
 }
